# docs/swagger.yaml
definitions:
...
  personalfinancedss_internal_module_identify_auth_dto.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
  personalfinancedss_internal_module_identify_auth_dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  personalfinancedss_internal_module_identify_auth_dto.ResetPasswordRequest:
    properties:
      new_password:
        maxLength: 128
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  personalfinancedss_internal_shared.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      error:
        type: string
      message:
        type: string
    type: object
  personalfinancedss_internal_shared.Success:
    properties:
      detail:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact:
    email: support@personalfinancedss.com
    name: API Support
    url: http://www.swagger.io/support
  description: Personal Finance Decision Support System API for managing accounts,
    transactions, budgets, goals, and investments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Personal Finance DSS API
  version: "1.0"
paths:
...
  /api/v1/auth/change-password:
    post:
      consumes:
      - application/json
      description: Change the password of the currently authenticated user
      parameters:
      - description: Password change data
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/personalfinancedss_internal_module_identify_auth_dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request a password reset link to be sent to email
      parameters:
      - description: Email address
        in: body
        name: forgot
        required: true
        schema:
          $ref: '#/definitions/personalfinancedss_internal_module_identify_auth_dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      summary: Request password reset
      tags:
      - auth
...
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password with reset token
      parameters:
      - description: Reset token and new password
        in: body
        name: reset
        required: true
        schema:
          $ref: '#/definitions/personalfinancedss_internal_module_identify_auth_dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/personalfinancedss_internal_shared.ErrorResponse'
      summary: Reset password
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: JWT Authorization header using the Bearer scheme. Enter 'Bearer'
      [space] and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"